backend/requirements.txt
Django==3.2.16
djangorestframework==3.14.0
djoser==2.2.2
gunicorn==20.1.0
psycopg2-binary==2.9.3
Pillow==9.0.0
django-colorfield==0.11.0
django-filter==23.3
python-dotenv==1.0.0
PyYAML==6.0
drf-extra-fields==3.7.0
django-cors-headers==3.13.0